+++ src/boinor/twobody/orbit/creation.py
@@ -74,7 +74,7 @@ class OrbitCreationMixin:
         if v.ndim != 1:
             raise ValueError(f"Vectors must have dimension 1, got {v.ndim}")
 
-        self._state = RVState(self.attractor, (self.r, self.v + v), self.plane)
+        self.state = RVState(self.attractor, (self.r, self.v + v), self.plane)
 
     @classmethod
     def from_coords(cls, attractor, coord, plane=Planes.EARTH_EQUATOR):